updates combat pages

+++ ModularTegustation/tegu_items/!combat_page/combat_page.dm
@@ -7,8 +7,8 @@
 	slot_flags = ITEM_SLOT_POCKETS
 	w_class = WEIGHT_CLASS_SMALL
 	var/combat_level = 1
-	var/reward_type = "PE"
-	var/reward_specification		//Any additional information.
+	var/reward_pe = 0
+	var/reward_items = list()		//Any additional information.
 	var/spawn_enemies = list(/mob/living/simple_animal/hostile/ordeal/amber_bug)
 	var/spawn_type = "all"			//All spawns all enemies, random spawns a random one.
 	var/spawn_number = 1
@@ -23,9 +23,13 @@
 		if(!istype(user) || !(user?.mind?.assigned_role in allowedroles))
 			to_chat(user, span_notice("The page glows red. It is unable to be used by you. Only Departmental officers can use this page."))
 			return
+
+	var/start_page = alert("Start this combat page?", "Combat Page", "Yes", "No")
+	if(start_page == "No")
+		return
 	Spawn_Reward(user)
 	Spawn_Combat(user)
-	minor_announce("A combat page has been started by [user.name]. Reward type: [reward_type]" , "[name]")
+	minor_announce("A combat page has been started by [user.name]." , "[name]")
 	qdel(src)
 
 /obj/item/combat_page/examine(mob/user)
@@ -34,11 +38,9 @@
 		. += span_notice("[special]")
 	. += span_notice("This is a level [combat_level] combat page")
 
-	switch(reward_type)
-		if("Item")
-			. += span_notice("You will recieve an item from this page")
-		if("PE")
-			. += span_notice("You will recieve PE from this page")
+	if(LAZYLEN(reward_items))
+		. += span_notice("You will recieve an item from this page")
+	. += span_notice("You will recieve [reward_pe] PE from this page")
 
 /obj/item/combat_page/proc/Spawn_Combat(mob/living/user)
 	if(!LAZYLEN(GLOB.xeno_spawn))
@@ -60,13 +62,9 @@
 		Spawn_Combat(user)
 
 /obj/item/combat_page/proc/Spawn_Reward(mob/living/user)
-	switch(reward_type)
-		if("Item")
-			if(!islist(reward_specification))				//Support for lists too!
-				new reward_specification(get_turf(user))	//Item uses the reward specification as an itempath
-				return
-			for(var/obj/I in reward_specification)
-				new I(get_turf(user))
+	if(LAZYLEN(reward_items))
+		for(var/obj/I in (reward_items))
+			new I(get_turf(user))
 
-		if("PE")
-			SSlobotomy_corp.AdjustAvailableBoxes(-1 * reward_specification)	//PE uses the reward specification as a number
+	if("reward_pe")
+		SSlobotomy_corp.AdjustAvailableBoxes(-1 * reward_pe)
